+++ src/main/java/net/rotgruengelb/quirl/mechanics/interact/shovel/v1/CustomShovelBlockInteract.java
@@ -1,9 +1,9 @@
-package net.rotgruengelb.quirl.behavior.shovel.v1;
+package net.rotgruengelb.quirl.mechanics.interact.shovel.v1;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
-import net.rotgruengelb.quirl.behavior.result.v1.ToolBlockInteractResult;
+import net.rotgruengelb.quirl.mechanics.interact.result.v1.ToolBlockInteractResult;
 
 import java.util.HashMap;
 import java.util.Map;
